rename package to invoicestubbed, as we are using stubs, and not mocks there
