+++ src/main/java/tudelft/invoicestubbed/InvoiceFilter.java
@@ -1,6 +1,5 @@
-package tudelft.invoicemocked;
+package tudelft.invoicestubbed;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import static java.util.stream.Collectors.toList;
